Check if coords are equal improved
Now we implement our own function to check if coordinates are
equal. When converting coords forth and back, rounding errors might lead
to situations, where the float values of coordinates are not equal.
Therefore now a comparison of coordinates is always done based on the
[N/S|E/W] <grad> <minutes.xxx> format.

Additionally within an svg no font-specific character should be used, as
the font might not be installed on the target host. Therefore all text
should be converted to graphs.